fixed issue with large range of values(height was 0)

+++ Viper/Sort/Animate.py
@@ -7,7 +7,8 @@ class Animate:
     def render(list,vip=[],vvip=[]):
         Window.instance.screen.fill((0,0,0))
         width = int(st.screen_width / (len(list))) -1
-        height = int(st.screen_height / (st.range)) 
+        height = st.screen_height / (st.range)
+        
         if len(list)/st.screen_width > 0.50:
             print("list is too long, go for something like",int(st.screen_width*0.5))
             pygame.quit()
@@ -17,7 +18,7 @@ class Animate:
                 color = (255,0,0)
             if index in vvip:
                 color = (39, 26, 232)
-            pygame.draw.rect(Window.instance.screen,color,((1+width) * index,(st.screen_height - bar * height)+1,width,bar*height))
+            pygame.draw.rect(Window.instance.screen,color,((1+width) * index,int(st.screen_height - bar * height)+1,width,int(bar*height)))
         Window.update()
     def start():    
         print(chr(27)+ "[2J")
